refactor: separate Moment from TableBody
